Django/requirements.txt
Django>=4.2
gunicorn>=20.1.0
psycopg2-binary>=2.9.6
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
celery>=5.2.7
aiogram>=2.23
psycopg2-binary>=2.9.6

